move get_database_yml
